Redefine `intersection` and `union` functions in `MultiSet` example.
Redefine these functions in terms of `MonoidMap.intersection` and
`MonoidMap.union` respectively.

Although the `MultiSet.intersection` function can be implemented in
terms of `gcd` from the `GCDMonoid` type class (which is defined in
terms of `min` for the `Sum Natural` type), there is currently no
`LCMMonoid` type class available, so we have no access to a general
`lcm` method for pairs of monoidal values.

Instead we simply apply `min` or `max` to matching multiplicities,
which gives us the desired behaviour for multisets:

```
> import qualified Data.Total.MonoidMap.Examples.MultiSet as MS

> MS.fromList "aaabbc" `MS.union` MS.fromList "abbccc"
"aaabbccc"

> MS.fromList "aaabbc" `MS.intersection` MS.fromList "abbccc"
"abbc"
```